Reorganize some directories
